feat: fix type

+++ internal/db/models.go
@@ -74,7 +74,7 @@ type Account struct {
 	User     string   `gorm:"not null"              json:"user"`
 	Account  uint64   `gorm:"not null"              json:"account"`
 	Chain    uint8    `gorm:"not null"              json:"chain"`
-	Index    uint64   `gorm:"not null"              json:"index"`
+	Index    uint32   `gorm:"not null"              json:"index"`
 	Address  string   `gorm:"uniqueIndex; not null" json:"address"`
 	LogIndex LogIndex `gorm:"foreignKey:ForeignID"` // has one https://gorm.io/zh_CN/docs/has_one.html
 }
